+++ internal/service/frontend/dag/response/dag_detail.go
