+++ internal/service/frontend/dag/response/logs.go
